requirements.txt
feedparser
google-genai
python-dotenv
sentence-transformers
newspaper3k
requests
lxml_html_clean
yt-dlp
snscrape
